requirements.txt
email-validator>=2.2.0
flask-login>=0.6.3
flask>=3.1.0
flask-sqlalchemy>=3.1.1
gunicorn>=23.0.0
psycopg2-binary>=2.9.10
python-barcode>=0.15.1
openpyxl>=3.1.5
werkzeug>=3.1.3
reportlab>=4.4.0
pandas>=2.2.3
pillow>=11.2.1
barcode>=1.0.4
sqlalchemy>=2.0.40
